Organise laser shooting functions
