--- src/freeFunctions/game/fireGunshipLaser.cpp
@@ -1,6 +1,0 @@
-#include "../../../include/models/iGunship.hpp"
-
-void fireGunshipLaser(IGunship &gunship)
-{
-  gunship.fire();
-}
